flask_api/src/requirements.txt
Flask==2.0.1
google-cloud-dialogflow-cx==1.3.2
